Do not reset on the last epoch

+++ testTwoPlayers.py
@@ -78,7 +78,7 @@ async def main(player1: Player, player2: Player, nTeams: int) -> None:
             sum(battle.turn for battle in player.battles.values()) / len(player.battles)
         )
 
-        if (epoch + 1) % 100 == 0:
+        if (epoch + 1) % 100 == 0 and (epoch + 1) != nEpochs:
             os.killpg(os.getpgid(p.pid), signal.SIGINT)
             p.wait()
             p = serverControl.startServer()
